requirements.txt
# TaskMaster - Project Requirements
# Recommended Python Version: 3.14+ (Compatible with Django 6.0.3)

Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.8
django-cors-headers==4.9.0
PyJWT==2.11.0
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
